+++ src/PanelTraits/FakeFields.php
@@ -14,20 +14,12 @@ trait FakeFields
      *
      * @return array
      */
-    public function compactFakeFields($request, $form = 'create')
+    public function compactFakeFields($request, $form = 'create', $id = false)
     {
         $fake_field_columns_to_encode = [];
 
         // get the right fields according to the form type (create/update)
-        switch (strtolower($form)) {
-            case 'update':
-                $fields = $this->update_fields;
-                break;
-
-            default:
-                $fields = $this->create_fields;
-                break;
-        }
+        $fields = $this->getFields($form, $id);
 
         // go through each defined field
         foreach ($fields as $k => $field) {