+++ CHANGELOG.md
@@ -40,11 +40,19 @@ All Notable changes to `Backpack CRUD` will be documented in this file
 - excluded _method from massAssignment, so create/update errors will be more useful;
 
 
+## [3.1.48] - 2016-12-14
+
+### Fixed
+- Prevent double-json-encoding on complicated field types, when using attribute casting; Fixes #259;
+
+
 ## [3.1.47] - 2016-12-14
 
 ### Fixed
 - Don't mutate date/datetime if they are empty. It will default to now;
 - select_from_array has a new option: "allows_multiple";
+- syncPivot is now done before saving the main entity in Update::edit();
+- added beforeColumn(), afterColumn(), beforeField() and afterField() methods to more easily reorder fields and columns - big up to [Ben Sutter](https://github.com/b8ne) for this feature;
 
 
 ## [3.1.46] - 2016-12-13